requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
astroid==2.4.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blessed==1.20.0
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.4
comtypes==1.1.7
Django==4.2.8
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.3
inflection==0.5.1
isort==5.6.4
jinxed==1.2.1
lazy-object-proxy==1.4.3
lxml==4.9.1
mccabe==0.6.1
msedge-selenium-tools==3.141.3
numpy==1.24.4
openpyxl==3.0.10
pandas==2.0.3
pylint==2.6.0
PyMySQL==1.0.2
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.9.0.post0
pytz==2020.4
pywin32==300
pywinauto==0.6.8
redis==3.5.3
requests==2.28.1
selenium==3.141.0
six==1.15.0
slacker==0.14.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.31
sqlparse==0.4.4
toml==0.10.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.6
wcwidth==0.2.13
wrapt==1.12.1
